test: move regex tests to diff_tests.rs
